+++ security/provenance/type.c
@@ -59,8 +59,10 @@ static const char RL_STR_PROC_READ[]             = "proc_read";
 static const char RL_STR_MMAP_EXEC[]             = "mmap_exec";
 static const char RL_STR_SND[]                   = "send";
 static const char RL_STR_SND_PACKET[]            = "send_packet";
+static const char RL_STR_SND_UNIX[]            	 = "send_unix";
 static const char RL_STR_RCV[]                   = "receive";
 static const char RL_STR_RCV_PACKET[]            = "receive_packet";
+static const char RL_STR_RCV_UNIX[]            	 = "receive_unix";
 static const char RL_STR_PERM_READ[]             = "perm_read";
 static const char RL_STR_PERM_WRITE[]            = "perm_write";
 static const char RL_STR_PERM_EXEC[]             = "perm_exec";
@@ -184,10 +186,14 @@ const char* relation_str(uint64_t type)
 		return RL_STR_SND;
 	case RL_SND_PACKET:
 		return RL_STR_SND_PACKET;
+	case RL_SND_UNIX:
+		return RL_STR_SND_UNIX;
 	case RL_RCV:
 		return RL_STR_RCV;
 	case RL_RCV_PACKET:
 		return RL_STR_RCV_PACKET;
+	case RL_RCV_UNIX:
+		return RL_STR_RCV_UNIX;
 	case RL_PERM_READ:
 		return RL_STR_PERM_READ;
 	case RL_PERM_WRITE:
@@ -261,6 +267,7 @@ uint64_t relation_id(const char* str)
 	MATCH_AND_RETURN(str, RL_STR_MMAP_EXEC, RL_MMAP_EXEC);
 	MATCH_AND_RETURN(str, RL_STR_SND, RL_SND);
 	MATCH_AND_RETURN(str, RL_STR_SND_PACKET, RL_SND_PACKET);
+	MATCH_AND_RETURN(str, RL_STR_SND_UNIX, RL_SND_UNIX);
 	MATCH_AND_RETURN(str, RL_STR_RCV, RL_RCV);
 	MATCH_AND_RETURN(str, RL_STR_RCV_PACKET, RL_RCV_PACKET);
 	MATCH_AND_RETURN(str, RL_STR_PERM_READ, RL_PERM_READ);